hkconv p1

+++ pcdet/models/__init__.py
@@ -39,7 +39,7 @@ def load_data_to_gpu(batch_dict):
         elif key in ['image_shape']:
             batch_dict[key] = torch.from_numpy(val).int().cuda()
         else:
-            batch_dict[key] = torch.from_numpy(val).float().cuda()
+            batch_dict[key] = torch.from_numpy(val).cuda()
 
 
 def model_fn_decorator():
